backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
httpx==0.27.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-multipart==0.0.9
qdrant-client==1.9.1
sentence-transformers==3.0.1
langchain-core>=0.2.19
langgraph>=0.1.9
langchain-openai>=0.1.0
